requirements.txt
Flask
flask-cors
pymongo
deepface
opencv-python
numpy
Pillow
